requirements.txt
biopython
cython
gunicorn
Flask
matplotlib
pandas
pytest
requests
viennarna